System system update
Created the system to use the users on the database to change the login system.
Change the authentication method to look the users/password in the Database.
Remove the scanner authenticate (in file user/password) and added the database connection scanner (checking user/password) directly in the database.
Added option to Admin role be able to create users and assign roles.
Changed the options (choices) menu in order to better visualization and structure.
Added the option for each user/role be able to change their own username and password.