Student report creation
Creating Student Report Generator:

Generating a report based on searching in the database the Student name and student number, the program that the student is undertaking, the modules the student is currently enrolled in, the modules that the student has completed, and the corresponding grades and the modules that the student needs to repeat.

This student report can be listed only by a person with the Office role.

User can print the report on screen.

Updated the Main.java in order to call the method generateStudentReport present in the OfficeReportGenerator (new file).

Checked/translated comments to EN.