Course report creation
Creating Course Report Generator:

Generating a report based on searching in the database the module name and the programme that module is in, the number of students enrolled in that module, the lecturer of that module and the room the module is assigned to classroom (or “online” if it is taught online).

This Course report can be listed only by a person with the Office role.

User can print the report on screen.

Updated the Main.java in order to call the method generateCourseReport present in the CourseReportGenerator (new file).

Minor changes added.